backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.1
boto3==1.35.16
botocore==1.35.16
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.0.8
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
jmespath==1.0.1
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-environ==0.4.54
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
